+++ imod/common/utilities/mask.py
@@ -1,6 +1,5 @@
 import numbers
 
-import numpy as np
 import xarray as xr
 from plum import Dispatcher
 from xarray.core.utils import is_scalar
@@ -11,9 +10,11 @@ from imod.common.interfaces.ipackage import IPackage
 from imod.common.interfaces.isimulation import ISimulation
 from imod.typing.grid import (
     GridDataArray,
+    concat,
     get_spatial_dimension_names,
     is_same_domain,
     is_spatial_grid,
+    notnull,
 )
 
 # create dispatcher instance to limit scope of typedispatching
@@ -142,15 +143,34 @@ def _adjust_mask_for_unlayered_data(
     return array_mask
 
 
-def mask_arrays(arrays: dict[str, xr.DataArray]) -> dict[str, xr.DataArray]:
+def mask_arrays(arrays: dict[str, GridDataArray]) -> dict[str, GridDataArray]:
     """
     This function takes a dictionary of xr.DataArrays. The arrays are assumed to have the same
     coordinates. When a np.nan value is found in any array, the other arrays are also
     set to np.nan at the same coordinates.
     """
-    masks = [xr.DataArray(~np.isnan(array)) for array in arrays.values()]
+    masks = [notnull(array) for array in arrays.values()]
     # Get total mask across all arrays
-    total_mask = xr.concat(masks[:], dim="arrays").all("arrays")
+    total_mask = concat(masks, dim="arrays").all("arrays")
     # Mask arrays with total mask
     arrays_masked = {key: array.where(total_mask) for key, array in arrays.items()}
     return arrays_masked
+
+
+def broadcast_and_mask_arrays(
+    arrays: dict[str, xr.DataArray],
+) -> dict[str, xr.DataArray]:
+    """
+    This function takes a dictionary of xr.DataArrays and broadcasts them to the same shape.
+    It then masks the arrays with np.nan values where any of the arrays have np.nan values.
+    """
+    # Broadcast arrays to the same shape
+    broadcasted_arrays = xr.broadcast(*arrays.values())
+    # Test if there is a spatial grid in the arrays, if not the broadcasting
+    # will result in no spatial grid.
+    if not is_spatial_grid(broadcasted_arrays[0]):
+        raise ValueError("One or more arrays need to be a spatial grid.")
+    broadcasted_arrays = dict(zip(arrays.keys(), broadcasted_arrays))
+
+    # Mask arrays with np.nan values
+    return mask_arrays(broadcasted_arrays)