+++ imod/mf6/ghb.py
@@ -5,6 +5,7 @@ from typing import Optional
 import numpy as np
 
 from imod.common.interfaces.iregridpackage import IRegridPackage
+from imod.common.utilities.mask import broadcast_and_mask_arrays
 from imod.common.utilities.regrid import _regrid_package_data
 from imod.common.utilities.regrid_method_type import RegridMethodType
 from imod.logging import init_log_decorator, standard_log_decorator
@@ -321,14 +322,14 @@ class GeneralHeadBoundary(BoundaryCondition, IRegridPackage):
             "head": imod5_data[key]["head"],
             "conductance": imod5_data[key]["conductance"],
         }
-        is_planar = is_planar_grid(data["conductance"])
-
         if regridder_types is None:
             regridder_types = GeneralHeadBoundaryRegridMethod()
 
         regridded_package_data = _regrid_package_data(
             data, idomain, regridder_types, regrid_cache, {}
         )
+        regridded_package_data = broadcast_and_mask_arrays(regridded_package_data)
+        is_planar = is_planar_grid(regridded_package_data["conductance"])
         if is_planar:
             layered_data = cls.allocate_and_distribute_planar_data(
                 regridded_package_data,